pyproject.toml
[project]
name = "nce-english-practices"
version = "0.1.0"
description = "English Tense Practice Web App"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.4",
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.53.0",
    "cryptography>=46.0.3",
    "sqlalchemy[asyncio]>=2.0.44",
    "asyncpg>=0.31.0",
    "alembic>=1.17.2",
    "pydantic-settings>=2.12.0",
    "aiosqlite>=0.21.0",
    "pyyaml>=6.0.3",
    "edge-tts>=6.1.10",
    "python-multipart>=0.0.9",
    "dashscope>=1.20.0",
    "jsonpatch>=1.33",
    "feedparser>=6.0.12",
    "ebooklib>=0.20",
    "lxml>=6.0.2",
]

[project.optional-dependencies]
dictionary = [
    "readmdict>=0.1.1",
    "python-lzo>=1.15",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "nest-asyncio>=1.6.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-env>=1.2.0",
    "pytest-playwright>=0.7.2",
    "ruff>=0.14.11",
]
